spilt out "run on sim" part to other crate `playdate-simulator-utils`
