+++ Cargo.toml
@@ -27,9 +27,10 @@ system = { version = "0.3", path = "api/system", package = "playdate-system", de
 sys = { version = "0.3", path = "api/sys", package = "playdate-sys", default-features = false }
 
 tool = { version = "0.1", path = "support/tool", package = "playdate-tool" }
-device = { version = "0.1", path = "support/device", package = "playdate-device" }
 build = { version = "0.2", path = "support/build", package = "playdate-build", default-features = false }
 utils = { version = "0.1", path = "support/utils", package = "playdate-build-utils", default-features = false }
+device = { version = "0.1", path = "support/device", package = "playdate-device" }
+simulator = { version = "0.1", path = "support/sim-ctrl", package = "playdate-simulator-utils", default-features = false }
 bindgen = { version = "0.1", path = "support/bindgen", package = "playdate-bindgen", default-features = false }
 bindgen-cfg = { version = "0.1", path = "support/bindgen-cfg", package = "playdate-bindgen-cfg", default-features = false }
 
@@ -46,3 +47,4 @@ serde = "1.0"
 serde_json = "1.0"
 toml = "0.8"
 futures-lite = "2.3"
+thiserror = "1.0"
